refactor(3c): extract BookingNotificationService from NotificationService
Pragmatic 2-interface split along the clean consumer seam — the background
NotificationWorker and the synchronous NotificationsController never called
the same methods, so the split needs no consumer to inject two interfaces.

New IBookingNotificationService owns the 3 background-only methods
(NotifyBookingCreated/Cancelled, CheckAndNotifyCapacity) plus the private
SendPushAsync (WebPush delivery, stale-sub pruning, push-outcome mirroring).
INotificationService slimmed 13 -> 10 members (CRUD/query + push-subscribe +
VAPID); NotificationService dropped 332 -> 115 lines and shed its
ITableRepository/IBookingRepository deps. NotificationWorker now resolves
IBookingNotificationService. NotificationsController unchanged.

8 booking/capacity tests moved to BookingNotificationServiceTests;
NotificationWorkerTests repointed to mock IBookingNotificationService.
Full suite: 917 pass / 2 pre-existing fail (parity).
