+++ OpenRestoApi/Infrastructure/Persistence/Repositories/TableRepository.cs
@@ -9,6 +9,7 @@ namespace OpenRestoApi.Infrastructure.Persistence.Repositories;
 [OnlyAccessibleBy("OpenRestoApi.Tests.Services.BookingServiceTests")]
 [OnlyAccessibleBy("OpenRestoApi.Tests.Services.AdminServiceTests")]
 [OnlyAccessibleBy("OpenRestoApi.Tests.Services.NotificationServiceTests")]
+[OnlyAccessibleBy("OpenRestoApi.Tests.Services.BookingNotificationServiceTests")]
 [OnlyAccessibleBy("OpenRestoApi.Tests.Services.RestaurantManagementServiceTests")]
 [OnlyAccessibleBy("OpenRestoApi.Tests.Services.WalkInTests")]
 [OnlyAccessibleBy("OpenRestoApi.Tests.Controllers.AdminControllerRestoreTests")]
